ElvenGard vdev HeavensStrike.Game.PacketResolver behaviour

Documentation for HeavensStrike.Game.LoginServer.

Functions

__using__(opts) (macro)

Use HeavensStrike.Game.PacketResolver behaviour.

Callbacks

deserialize(data)

deserialize(data :: binary()) :: list()

Transform a raw packet to an understandable packet. You can, for example, apply your cryptographic algorithm and split your packet. This function must return a list starting with your packet header followed by params.

resolve(client, data)

resolve(client :: HeavensStrike.Game.Client.t(), data :: binary()) ::
  {:cont}
  | {:halt, {:ok, term()}}
  | {:halt, {:error, HeavensStrike.Game.LoginServer.conn_error()}}

Just split a packet and call his packet handler