ElvenGard vdev LoginServer.PacketResolver

Parse a packet and call the right PacketHandler

Functions

deserialize(data)

Transform a raw packet to an understandable packet. You can, for example, apply your cryptographic algorithm and split your packet. This function must return a list starting with your packet header followed by params.

Callback implementation for HeavensStrike.Game.PacketResolver.deserialize/1.

resolve(client, data)

Just split a packet and call his packet handler

Callback implementation for HeavensStrike.Game.PacketResolver.resolve/2.